+++ src/org/elixir_lang/structure_view/node_provider/Used.kt
@@ -15,7 +15,6 @@ import org.elixir_lang.psi.ElixirAccessExpression
 import org.elixir_lang.psi.QualifiableAlias
 import org.elixir_lang.psi.call.Call
 import org.elixir_lang.psi.impl.call.finalArguments
-import org.elixir_lang.psi.impl.call.macroChildCalls
 import org.elixir_lang.psi.impl.stripAccessExpression
 import org.elixir_lang.structure_view.element.*
 import org.elixir_lang.structure_view.element.modular.Module
@@ -103,7 +102,7 @@ class Used : FileStructureNodeProvider<TreeElement>, ActionShortcutProvider {
 
                                         if (org.elixir_lang.psi.Module.`is`(call)) {
                                             val module = Module(call)
-                                            val childCalls = call.macroChildCalls()
+                                            val childCalls = org.elixir_lang.psi.CallDefinitionClause.modularChildCalls(call).toTypedArray()
 
                                             val macroByNameArity = HashMap<NameArity, CallDefinition>(childCalls.size)