+++ testData/org/elixir_lang/heex/reference/heex_component/conditional_component/page_live.ex
@@ -1,0 +1,7 @@
+defmodule MyAppWeb.PageLive do
+  def button(assigns), do: assigns
+
+  if Code.ensure_loaded?(Kernel) do
+    def gated(assigns), do: assigns
+  end
+end